+++ lib/src/widgets/cells/table_plus_selection_cell.dart
@@ -43,6 +43,9 @@ class TablePlusSelectionCell extends StatelessWidget {
           color: Colors.transparent,
           child: InkWell(
             onTap: rowId != null ? () => onSelectionChanged(rowId!) : null,
+            mouseCursor: rowId != null
+                ? SystemMouseCursors.click
+                : MouseCursor.defer,
             child: content,
           ),
         );